pyproject.toml
[project]
name = "salesforce-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.3.0",
    "mcp-server>=0.1.4",
    "simple-salesforce>=1.12.9",
    "starlette>=0.48.0",
    "uvicorn>=0.38.0",
]
